Extract ChecksumTableStatement to reduce duplicate logic for MySQL and doris
